空気タンクアニメーション分割
