+++ src/game/entities/UIFactory.ts
@@ -1,17 +1,17 @@
 import { Entity } from '@core/ecs/entity'
 import { World } from '@core/ecs/world'
-import { playerUIAI } from '@game/ai/entity/playerUI/playerUIAI'
+import { hudAI } from '@game/ai/entity/hud/hudAI'
 import { AiComponent } from '@game/components/aiComponent'
 import { EntityFactory } from './entityFactory'
 
-export class PlayerUIFactory extends EntityFactory {
+export class UIFactory extends EntityFactory {
   public constructor(private world: World) {
     super()
   }
 
   public create(): Entity {
     const entity = new Entity()
-    entity.addComponent('Ai', new AiComponent(playerUIAI(this.world)))
+    entity.addComponent('Ai', new AiComponent(hudAI(this.world)))
 
     return entity
   }
